Add module_19_1

+++ module_19/UrbanDjango2/task1/models.py
@@ -14,7 +14,7 @@ class Buyer(models.Model):
 class Game(models.Model):
     title = models.CharField(max_length=30)
     cost = models.DecimalField(max_digits=10, decimal_places=2)
-    size = models.IntegerField()
+    size = models.DecimalField(max_digits=10, decimal_places=2)
     description = models.TextField()
     age_limited = models.BooleanField(default=False)
     buyer = models.ManyToManyField(Buyer)
